Make it work on Windows.
Changed the project and some source code to make the project work with
the windows Visual Studio version. It now compiles and runs but shows no
console debug information on the scene. I seem to have problems making
the Console work correctly. There might be a problem with the include
files or the linker configuration.